1.5.1 完成后台管理的章节的增删改查 视频管理的页面

+++ src/asia/lhweb/lhmooc/model/vo/CourseVo.java
@@ -38,7 +38,7 @@ public class CourseVo {
     /**
      * 课程价格
      */
-    private Integer price;
+    private Double price;
 
     /**
      * 课程封面的图片
@@ -207,17 +207,11 @@ public class CourseVo {
         this.profile = profile;
     }
 
-    /**
-     * 课程价格
-     */
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    /**
-     * 课程价格
-     */
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
